Update Python versions
I bumped all Python versions to the latest patch release. In addition,
I changed garnet's Python 3 version from 3.5.x to 3.6.x, since if it
is only going to install one of each of Python 2 and 3, then it makes
sense for it to be the latest of each.

For latest version numbers see:
https://www.python.org/downloads/