Remove stevonnie docker package version override
Once travis-ci/travis-cookbooks#892 is merged, the default docker-ce
package version (`17.06.1~ce-0~ubuntu`) exists for both Trusty and
Xenial, so doesn't need to be adjusted for stevonnie.

Should the Trusty and Xenial versions ever diverge again, it would be
best to add the conditional to travis-cookbooks rather than overriding
in packer-templates.